Move Ollama classifier to intermediate layer
