+++ src/adapters/adapter_ollama.hpp
@@ -12,10 +12,9 @@ namespace adapters {
 class Ollama: public CurlBase {
 public:
     Ollama(const Configurations &configs);
-    std::expected<OllamaClassification, OllamaError> classify_instructions(const std::string &prompt);
     std::expected<OllamaEdit, OllamaError> query_edit_code(const std::string &prompt);
 
-private:
+protected:
     std::string query_chat_api_(const std::string &post_fields);
 
     std::string model_;
